Move thumbnail array to global constants.
