+++ src/components/Sketches/index.js
@@ -1,17 +1,16 @@
 import '../../styles/Sketches.scss';
+import React from "react";
+import SketchBox from "./components/SketchBox";
+import ConfirmDeleteModalContainer from "./containers/ConfirmDeleteModalContainer";
+import CreateSketchModalContainer from "./containers/CreateSketchModalContainer";
+import EditSketchModalContainer from "./containers/EditSketchModalContainer";
+import OpenPanelButtonContainer from "../common/containers/OpenPanelButtonContainer";
+import { ThumbnailArray } from "../../constants";
+import CodeDownloader from "../../util/languages/CodeDownloader";
 
 import { faFile } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
-import React from 'react';
-import { Button } from 'reactstrap';
-import CodeDownloader from '../../util/languages/CodeDownloader';
-import OpenPanelButtonContainer from '../common/containers/OpenPanelButtonContainer';
-import SketchBox from './components/SketchBox';
-import { SketchThumbnailArray } from './constants';
-import ConfirmDeleteModalContainer from './containers/ConfirmDeleteModalContainer';
-import CreateSketchModalContainer from './containers/CreateSketchModalContainer';
-import EditSketchModalContainer from './containers/EditSketchModalContainer';
 
 const ROW_PADDING = 100;
 const SKETCH_WIDTH = 220;
@@ -30,7 +29,9 @@ class Sketches extends React.Component {
     };
   }
 
-  getRandomSketchThumbnail = () => SketchThumbnailArray[Math.floor(Math.random() * SketchThumbnailArray.length)];
+  getRandomSketchThumbnail = () => {
+    return ThumbnailArray[Math.floor(Math.random() * ThumbnailArray.length)];
+  };
 
   setCreateSketchModalOpen = (val) => {
     this.setState({ createSketchModalOpen: val });
@@ -72,10 +73,10 @@ class Sketches extends React.Component {
   );
 
   getThumbnailSrc = (val) => {
-    if (val === undefined || val === '' || val >= SketchThumbnailArray.length || val < 0) {
-      return SketchThumbnailArray[0];
+    if (val === undefined || val === "" || val >= ThumbnailArray.length || val < 0) {
+      return ThumbnailArray[0];
     }
-    return SketchThumbnailArray[val];
+    return ThumbnailArray[val];
   };
 
   renderSketches = () => {
