+++ src/constants/index.js
@@ -44,6 +44,67 @@ const OUTPUT_ONLY = 2;
 // UI
 const EDITOR_WIDTH_BREAKPOINT = 1000;
 
+const ThumbnailArray = [
+  "Ant",
+  "Badger",
+  "Bear",
+  "Beaver",
+  "Bird",
+  "Bug",
+  "Bull",
+  "Bumblebee",
+  "Butterfly",
+  "Cat",
+  "Caterpillar",
+  "Chicken",
+  "Clown Fish",
+  "Corgi",
+  "Cow",
+  "Crab",
+  "Deer",
+  "Dinosaur",
+  "Dog",
+  "Dolphin",
+  "Dragonfly",
+  "Duck",
+  "Elephant",
+  "Falcon",
+  "Fish",
+  "Fly",
+  "Frog",
+  "Giraffe",
+  "Gorilla",
+  "Grasshopper",
+  "Horse",
+  "Hummingbird",
+  "Insect",
+  "Kangaroo",
+  "Kiwi Bird",
+  "Ladybird",
+  "Leopard",
+  "Lion",
+  "Llama",
+  "Mite",
+  "Mosquito",
+  "Octopus",
+  "Panda",
+  "Prawn",
+  "Puffin Bird",
+  "Rabbit",
+  "Rhinoceros",
+  "Seahorse",
+  "Shark",
+  "Sheep",
+  "Snail",
+  "Spider",
+  "Starfish",
+  "Stork",
+  "Turtle",
+  "Unicorn",
+  "Whale",
+  "Wolf",
+];
+
 module.exports = {
   // photo names
   PHOTO_NAMES,
@@ -82,4 +143,6 @@ module.exports = {
   // Firebase constants
   EMAIL_DOMAIN_NAME: '@fake.com',
   PANEL_IMAGE_SELECTOR_SIZE,
+
+  ThumbnailArray,
 };
